Clear calculator if number pressed after equals
Previously if a number  was pressed straight after equals, it was using the old
operator from the last operation and setting the number pressed as the second
number.

Now, it will instead clear both the first number and the operator and restart
if a number is presssed straight after equals.